foooooo

+++ RaindropOperation.cpp
@@ -48,7 +48,7 @@ RaindropOperation::RaindropOperation(VariableStore
 void RaindropOperation::hitRaindrop(Raindrop &drop) {
 
     const int saturation = random_in_range(saturation_min.getFloat() * 1000, saturation_max.getFloat() * 1000);
-    drop.color.hue = random_in_range(hue_min.getInteger(), hue_max.getInteger());
+    drop.color.hue = random_in_range(hue_min.getFloat(), hue_max.getFloat());
     drop.color.saturation = float(saturation) / 1000.0f;
     drop.color.value = float(random_in_range(value_min.getFloat() * 10000, value_max.getFloat() * 10000)) / 10000.0f;
